requirements.txt
matplotlib==3.7.1
numpy==1.24.1
open_clip_torch==2.20.0
opencv_contrib_python_headless==4.7.0.72
Pillow==9.3.0
Pillow==9.5.0
pycocotools==2.0.6
segment_anything==1.0
skimage==0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
